Update 3.16.83op.config
Hides the cursor by default, given that we are now using plymouth for boot screens. This gives a clean startup when using image-only boot screens and our default, 'tribar', re-enables the cursor upon completion, which is perfect. So to 'details' and 'text'.

However, if the 'nosplash' kernel parameter is used, the cursor won't be present on the text login screen.